完成基本的网关配置界面,后续细节待完善

+++ src/MicroService.ApiGateway.Application/Ocelot/IReRouteAppService.cs
@@ -14,7 +14,7 @@ namespace MicroService.ApiGateway.Ocelot
 
         Task<ReRouteDto> GetByRouteNameAsync(string routeName);
 
-        Task<ReRouteDto> GetAsync(int routeId);
+        Task<ReRouteDto> GetAsync(long routeId);
 
         Task<ReRouteDto> CreateAsync(ReRouteDto routeDto);
